Test fix (#247)
* Add fix to claim nft issue

* Remove toString()

* Remove uneeded exclaimation mark

* Update code to assume merkle amount is in proper hex value

* Update merkle file

+++ nft-drop/src/hooks/useFetchNFTDropData.ts
@@ -37,7 +37,7 @@ const useFetchNFTDropData = (
       return;
     }
 
-    const tokenId = parseFloat(BigNumber.from(merkleObj.amount).toString());
+    const tokenId = parseInt(BigNumber.from(merkleObj.amount).toString());
     const tokenIdOwner = await contract.ownerOf(tokenId).catch(() => undefined);
 
     setNFTDropData({
